# Add constexpr to improve performance