--- a/tests/unittests/const_unittest.cpp
+++ b/tests/unittests/const_unittest.cpp
@@ -31,8 +31,9 @@ TEST(CONST, PlayerGetNameLength) {
     EXPECT_GE(MAX_PLAYER_GET_NAME_LENGTH, std::strlen(p.getName()));
     for (u_int8_t i = 0; i <= MAX_PLAYER_NAME_LENGTH; i++) {
         CheckPlayer p{std::string(i, 'a').c_str()};
-        p.setPlayerPosNum(-1);
-        EXPECT_GE(MAX_PLAYER_GET_NAME_LENGTH, std::strlen(p.getName()));
+        p.setPlayerPosNum(0);
+        // actual max is name255 but playerPosNum is limited by MAX_PLAYERS, test with 0 that is always valid (but we need to subtract 2 for the missing digits)
+        EXPECT_GE(MAX_PLAYER_GET_NAME_LENGTH-2, std::strlen(p.getName()));
     }
 }