Config(10000, 5, 1000, 10, 1)
Start: 6.05s - 6.1s
Config: 6s - 6.05s (-0.05s)
Deck: 5.95s-6.025s (-0.1s)
Enums: 5.95s-6s (-0.1s)